carousel state object
